+++ processor/src/main/java/arez/persist/processor/Constants.java
@@ -7,9 +7,13 @@ final class Constants
   @Nonnull
   static final String AREZ_COMPONENT_CLASSNAME = "arez.annotations.ArezComponent";
   @Nonnull
+  static final String ACT_AS_COMPONENT_CLASSNAME = "arez.annotations.ActAsComponent";
+  @Nonnull
   static final String OBSERVABLE_CLASSNAME = "arez.annotations.Observable";
   @Nonnull
   static final String PERSIST_TYPE_CLASSNAME = "arez.persist.PersistType";
+  @Nonnull
+  static final String PERSIST_ID_CLASSNAME = "arez.persist.PersistId";
   @Nonnull
   static final String PERSIST_CLASSNAME = "arez.persist.Persist";
   @Nonnull